feat: introduce mosquitto_calls
This enables access to publish on mosquitto without a reference to the
plugin itself. Useful when sending messages without reacting
to an event on mqtt

BREAKING CHANGE: To migrate switch "self.broker_broadcast_publish" to
"mosquitto_calls::publish_broadcast"
